Doing some refactoring to make this an official repo we can use.
Initial tests show this is basically what we we've been wanting for over a year now.  Model is returning multiple segmented classes per image with labels.

I am just cleaning up how everything is generated so we can deploy this remotely in the future.  I am also adding integration for ClearML so we can monitor training.