Found an issue where I did not have a large enough swap file, so added some notes for future me to remember how to solve this issue.
You'll know you hit it if during training your process dies for no reason, and you just see a `process killed` message at the end of the output.  That means your OS ran out of resources and chose to kill the process.

This documentation has a few solutions to make sure that does not happen.